# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Deep Rock Galactic",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/contrats/": {
      "get": {
        "tags":["Contrats"],
        "description": "Obtient les contrats enregistrés",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags":["Contrats"],
        "description": "Crée un contrat",
        "produces": [
          "application/json"
        ],
        "parameters": [
          
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "planeteId": {
                  "example": "6260366953cb2638e9e6ea12"
                },
                "prime": {
                  "example": 30
                },
                "danger": {
                  "example": 3
                },
                "ressource": {
                  "example": "Cuivre"
                },
                "quantiteRessource": {
                  "example": 28000
                },
                "dateExpiration": {
                  "example": "2022-09-03T20:42:13.449Z"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/planetes/": {
      "get": {
        "tags":["Planetes"],
        "description": "Obtient toutes les planètes exploitées par la compagnie",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags":["Planetes"],
        "description": "Crée une planète tel un dieu de la galaxie",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nom": {
                  "example": "Jarvan IV"
                },
                "image": {
                  "example": "4124790129dad.jpeg"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}